requirements.txt
PyYAML
pynng
requests
google-cloud
protobuf
flask